解决代码中的绝大部分bug，开始编写毕业论文。

+++ cloud-collect-core/src/main/kotlin/com/windea/demo/cloudcollect/core/service/impl/HistoryServiceImpl.kt
@@ -18,10 +18,10 @@ class HistoryServiceImpl(
 	override fun create(history: History) {
 		//相对于最后浏览记录，不要重复添加浏览记录
 		val lastHistory = historyRepository.findFirstByUserIdOrderByIdDesc(history.user.id)
-		if(history.collect.id == lastHistory.collect.id) return
+		if(lastHistory != null && history.collect.id == lastHistory.collect.id) return
 		
 		//首先要删除之前的重复的浏览记录
-		historyRepository.deleteByCollectIdAndUserId(history.collect.id, history.user.id)
+		if(lastHistory != null) historyRepository.deleteByCollectIdAndUserId(history.collect.id, history.user.id)
 		
 		historyRepository.save(history)
 	}
